Fix some warnings

+++ include/InputManager.hpp
@@ -31,17 +31,17 @@ namespace Trn {
 		std::tuple<double, double> getMouseDelta() const;
 		void setCursorMoveCallback(std::function<void(double, double, double, double)> callback);
 	private:
-		static constexpr float doublePressInterval = 0.5f;
+		static constexpr double doublePressInterval = 0.5f;
 
 		static void keyCallback(GLFWwindow* window, int key, int scancode, int action, int mods);
 		KeyState previousState[GLFW_KEY_LAST] = { KeyState::Released };
 		KeyState currentState[GLFW_KEY_LAST] = { KeyState::Released };
-		float lastTimePressed[GLFW_KEY_LAST] = { 0.0f };
+		double lastTimePressed[GLFW_KEY_LAST] = { 0.0f };
 
 		static void mouseButtonCallback(GLFWwindow* window, int button, int action, int mods);
 		KeyState prevMouseState[GLFW_MOUSE_BUTTON_LAST] = { KeyState::Released };
 		KeyState mouseState[GLFW_MOUSE_BUTTON_LAST] = { KeyState::Released };
-		float mouseLastTimePressed[GLFW_MOUSE_BUTTON_LAST] = { 0.0f };
+		double mouseLastTimePressed[GLFW_MOUSE_BUTTON_LAST] = { 0.0f };
 
 		static void cursorPosCallback(GLFWwindow* window, double xpos, double ypos);
 		static void cursorEnterCallback(GLFWwindow* window, int entered);
